feat: clean-up get_price_or_fail with addtl test
